+++ src/middleware/packages/crypto/keys/index.js
@@ -1,5 +1,5 @@
 module.exports = {
   KEY_TYPES: require('./keyTypes'),
   MigrationService: require('./migration-service'),
-  KeyService: require('./key-service')
+  KeysService: require('./keys-service')
 };
